# Improve Reflection Performance for Beans with Superclass methods/fields