Avoid NoSuchMethodException being thrown and having resultant fillInStackTrace impact reflection performance. Here is a stacktrace I see a lot when running a load test:

java.base@11.0.5/java.lang.Throwable.fillInStackTrace(Native Method)
java.base@11.0.5/java.lang.Throwable.fillInStackTrace(Throwable.java:787)
   - locked java.lang.NoSuchMethodException@7e48719e
java.base@11.0.5/java.lang.Throwable.<init>(Throwable.java:270)
java.base@11.0.5/java.lang.Exception.<init>(Exception.java:66)
java.base@11.0.5/java.lang.ReflectiveOperationException.<init>(ReflectiveOperationException.java:56)
java.base@11.0.5/java.lang.NoSuchMethodException.<init>(NoSuchMethodException.java:51)
java.base@11.0.5/java.lang.Class.getDeclaredMethod(Class.java:2475)
app//io.quarkus.arc.impl.Reflections.findMethod(Reflections.java:33)
app//guru.jini.services.property.rest.PropertiesResource_Subclass.<init>(PropertiesResource_Subclass.zig:522)
app//guru.jini.services.property.rest.PropertiesResource_Bean.create(PropertiesResource_Bean.zig:80)
app//guru.jini.services.property.rest.PropertiesResource_Bean.create(PropertiesResource_Bean.zig:648)
app//io.quarkus.arc.impl.RequestContext.get(RequestContext.java:67)
...
This PR Introduces a weak set cache to track methods and fields that are found on superclasses so as to avoid always getting NoSuchMethodException.

Load tests against this change show a 36% increase in throughput for a basic rest service where the service extends another class which has about 8 methods on it.

Although this change helps a lot, it would be even better if the generated code for beans that extend other classes don't use reflection on every superclass method/field for every single class instantiation. Perhaps someone who understands that part better could take a look at optimising treatment of beans with superclasses as they perform significantly worse than ones that dont extend anything.