@gsmet 2020-01-31
@bcluap any chance you could properly rebase this PR and have one commit with the changes?

I also have one question: you're talking about throughput improvements, is it at runtime when you access the REST service?