@mkouba 2020-01-31
So we only use reflection if an injection point resolves to a dependent bean that injects `InjectionPoint` metadata. A common example is injection of a `@ConfigProperty` that is satisifed by a `io.smallrye.config.inject.ConfigProducer` method. I like the idea of a cache but we could probably cache all lookups (hit, miss, etc.). I'll take a look at this next week...