@gsmet 2020-01-31
@bcluap any chance you could provide a simple app where we can see the difference? Might be useful to validate the improvements.

Thanks!