@bcluap 2020-02-01
Test results of the test project (after about 10 runs)

Without the cache:

paul@PCBLaptop:~$ ab -c 10 -n 100000 http://localhost:8080/hellofast
This is ApacheBench, Version 2.3 <$Revision: 1807734 $>
Copyright 1996 Adam Twiss, Zeus Technology Ltd, http://www.zeustech.net/
Licensed to The Apache Software Foundation, http://www.apache.org/

Benchmarking localhost (be patient)
Completed 10000 requests
Completed 20000 requests
Completed 30000 requests
Completed 40000 requests
Completed 50000 requests
Completed 60000 requests
Completed 70000 requests
Completed 80000 requests
Completed 90000 requests
Completed 100000 requests
Finished 100000 requests


Server Software:        
Server Hostname:        localhost
Server Port:            8080

Document Path:          /hellofast
Document Length:        2 bytes

Concurrency Level:      10
Time taken for tests:   15.189 seconds
Complete requests:      100000
Failed requests:        0
Total transferred:      8000000 bytes
HTML transferred:       200000 bytes
Requests per second:    6583.75 [#/sec] (mean)
Time per request:       1.519 [ms] (mean)
Time per request:       0.152 [ms] (mean, across all concurrent requests)
Transfer rate:          514.36 [Kbytes/sec] received

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:        0    0   0.4      0      20
Processing:     0    1   1.2      1      50
Waiting:        0    1   1.1      1      50
Total:          0    1   1.3      1      50

Percentage of the requests served within a certain time (ms)
  50%      1
  66%      1
  75%      2
  80%      2
  90%      2
  95%      3
  98%      5
  99%      6
 100%     50 (longest request)


With the PR applied:

paul@PCBLaptop:~$ ab -c 10 -n 100000 http://localhost:8080/hellofast
This is ApacheBench, Version 2.3 <$Revision: 1807734 $>
Copyright 1996 Adam Twiss, Zeus Technology Ltd, http://www.zeustech.net/
Licensed to The Apache Software Foundation, http://www.apache.org/

Benchmarking localhost (be patient)
Completed 10000 requests
Completed 20000 requests
Completed 30000 requests
Completed 40000 requests
Completed 50000 requests
Completed 60000 requests
Completed 70000 requests
Completed 80000 requests
Completed 90000 requests
Completed 100000 requests
Finished 100000 requests


Server Software:        
Server Hostname:        localhost
Server Port:            8080

Document Path:          /hellofast
Document Length:        2 bytes

Concurrency Level:      10
Time taken for tests:   12.265 seconds
Complete requests:      100000
Failed requests:        0
Total transferred:      8000000 bytes
HTML transferred:       200000 bytes
Requests per second:    8153.36 [#/sec] (mean)
Time per request:       1.226 [ms] (mean)
Time per request:       0.123 [ms] (mean, across all concurrent requests)
Transfer rate:          636.98 [Kbytes/sec] received

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:        0    0   0.2      0       8
Processing:     0    1   0.9      1      37
Waiting:        0    1   0.9      1      37
Total:          0    1   1.0      1      37

Percentage of the requests served within a certain time (ms)
  50%      1
  66%      1
  75%      1
  80%      1
  90%      2
  95%      2
  98%      3
  99%      5
 100%     37 (longest request)

And with @Transactional removed:

paul@PCBLaptop:~$ ab -c 10 -n 100000 http://localhost:8080/hellofast
This is ApacheBench, Version 2.3 <$Revision: 1807734 $>
Copyright 1996 Adam Twiss, Zeus Technology Ltd, http://www.zeustech.net/
Licensed to The Apache Software Foundation, http://www.apache.org/

Benchmarking localhost (be patient)
Completed 10000 requests
Completed 20000 requests
Completed 30000 requests
Completed 40000 requests
Completed 50000 requests
Completed 60000 requests
Completed 70000 requests
Completed 80000 requests
Completed 90000 requests
Completed 100000 requests
Finished 100000 requests


Server Software:        
Server Hostname:        localhost
Server Port:            8080

Document Path:          /hellofast
Document Length:        2 bytes

Concurrency Level:      10
Time taken for tests:   10.904 seconds
Complete requests:      100000
Failed requests:        0
Total transferred:      8000000 bytes
HTML transferred:       200000 bytes
Requests per second:    9171.35 [#/sec] (mean)
Time per request:       1.090 [ms] (mean)
Time per request:       0.109 [ms] (mean, across all concurrent requests)
Transfer rate:          716.51 [Kbytes/sec] received

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:        0    0   0.2      0      12
Processing:     0    1   0.8      1      19
Waiting:        0    1   0.7      0      19
Total:          0    1   0.8      1      19
WARNING: The median and mean for the waiting time are not within a normal deviation
        These results are probably not that reliable.

Percentage of the requests served within a certain time (ms)
  50%      1
  66%      1
  75%      1
  80%      1
  90%      2
  95%      2
  98%      3
  99%      4
 100%     19 (longest request)


P.S. Set as @ApplicationScoped or @Singleton with no transaction and results don't change much from those with @RequestScoped which is great.